Swap column order for intake-episode table
intake-episode is a one-to-one link table. Its required in two contexts:
* Hub
* Combined Intake and Hub

Both these contexts require episodes data to be included, but intake
data is only required for the Combined context.

Reordering so that the first two columns are the foreign keys to data
always required in the upload, making them more consistent with then
other files.