Update service contact start time validations
* missing time format of 24:00
* Can not be used for new records
* Records with valid times can not be set to missing.

Rationale:
Exports are always "in the current spec version", so those uploading
PMHC v2 will get PMHC v4 exports. If they attempt to upload that (aka
round-tripping), it would have failed with empty start times.

The XMLSchema spec permits 24:00 as a valid "end of day" time,
which is canonically equivilent to 00:00. We'll use this as our
"missing" string for start times to allow export/upload round-trip.

This required additional validation rules to limit usage; As start time
is a required value, we do not permit 24:00 on NEW records, nor do we
permit existing records to change a valid time to missing.